Normal Theory Based Confidence Envelope for Density

This function creates a normal theory based confidence enevelope for the empirical density. The idea came from functions written by Bowman and Azzalini.

This function creates a confidence envelope for the empirical density by bootstrapping from the data. Transparency is used to indicate the level of uncertainty.

Arguments

h
A normal kernel function is used and h is its standard deviation. If this parameter is omitted, a normal optimal smoothing parameter is used.

fill
Fill color for the confidence envelope. The default is fill="skyblue"

model
The model to draw the confidence envelope for. The default is model="none" which creates the confidence envelope from the data. Using model="normal" creates the confidence envelope based on a normal distribution.

k
Number of bootstrapped smoothers. The default is 1000.

alpha
Transparency level for the paths that make up the bootstrapped fitted lines. This may need to be adjusted if the argument k= is changed. The default value is alpha=0.03.

color
Color for the bootstrapped densities that make up the confidence envelope. The default is color="#1D91C0"

Aesthetics

These stat uses geom_ribbon() so support the same aesthetics: alpha, colour, linetype and size.