Watercolor Smoother Felix

This function creates a visually regression smoother, also called a Watercolor plot. The initial idea came from Solomon Hsiang. Felix Schonbrodt provided inital syntax based on Solomon's idea, and with additional inputfrom many blog commenters. He licensed this with a FreeBSD License (see syntax).

watercolor_smoother_felix(formula, data, title = "", B = 1000,
  shade = TRUE, shade.alpha = 0.1, spag = FALSE,
  spag.color = "darkblue", mweight = TRUE, show.lm = FALSE,
  show.median = TRUE, median.col = "white", show.points = FALSE,
  shape = 21, show.CI = FALSE, method = loess, bw = FALSE,
  slices = 200, palette = RColorBrewer::brewer.pal(9, "YlGnBu"),
  ylim = NULL, quantize = "continuous", add = FALSE, ...)

Arguments

B
Number of bootstrapped smoothers. The default is 1000.

shade
Logical value indicating whether to plot the shaded confidence region. The default is TRUE.

shade.alpha
Numercial value indicating the degree to which the confidence enevelope shading should fade out at the edges. This reduces the alpha value. (0 = no alpha decrease, 0.1 = medium alpha decrease, 0.5 = strong alpha decrease)

spag
Logical value indicating whether the bootstrapped spaghetti lines should be plotted. The default is FALSE.

spag.color
Color of bootstrapped spaghetti lines.

mweight
Logical value indicating whether the median smoother should be visually weighted. The default is TRUE.

show.lm
Logical value indicating whether the linear regresison line should be plotted. The default is FALSE.

show.median
Logical value indicating whether the median smoother should be plotted. The default is TRUE.

median.col
Color of the median smoother.

show.points
Logical value indicating whether to plot the observed data. The default is FALSE.

shape
Shape of points. The default is 21 (open circles). This is only relevant if show.points=TRUE.

show.CI
Logical value indicating whether the 95 envelope should be plotted. Default is FALSE.

method
the fitting function for the spaghettis; default: loess

bw
Logical value indicating whether to use a black-and-white palette. The default is FALSE.

slices
Number of slices in x and y direction for the shaded region. Higher numbers make a smoother plot, but takes longer to draw. I would not go beyond 500.

palette
Color palette for the shading in the confidence envelope. This defaults to the 9-color YlGnBu palette from RColorBrewer.

ylim
Set y-limit to restrict the range of the watercoloring. The default is NULL.

quantize:
Character string indicating the type of shading on the confidence envelope. The default is "continuous", which uses a continuous color gradient across the entire range of y. Setting this to "SD" we get three color regions for 1, 2, and 3 SD (an idea of John Mashey)

add:
Logical value indicating whether to return a new ggplot. The default is FALSE. If add = TRUE, only the elements are returned, which can be added to another ggplot (with the '+' operator)

...:
Further parameters passed to the fitting function, in the case of loess, for example, "span = .9", or "family = 'symmetric'"

Value

A ggplot object giving the watercolor plot